QMX working again

+++ source/query.h
@@ -35,8 +35,8 @@ namespace JASS
 		{
 		public:
 //			typedef uint32_t ACCUMULATOR_TYPE;									///< the type of an accumulator (probably a uint16_t)
-//			typedef uint16_t ACCUMULATOR_TYPE;									///< the type of an accumulator (probably a uint16_t)
-			typedef uint8_t ACCUMULATOR_TYPE;									///< the type of an accumulator (probably a uint16_t)
+			typedef uint16_t ACCUMULATOR_TYPE;									///< the type of an accumulator (probably a uint16_t)
+//			typedef uint8_t ACCUMULATOR_TYPE;									///< the type of an accumulator (probably a uint16_t)
 			typedef uint32_t DOCID_TYPE;										///< the type of a document id (from a compressor)
 
 		public:
